Update JAVACMD in universalJavaApplicationStub.sh
As Java 1.8 is old and no more supported, running this up on new new Java versions doesn't work. To fix the issue the user is forced to install Java 8 which override the existing new Java. With this new ENV var, we can install Java 1.8 in a dedicated folder without polluting our JAVA_HOME or our PATH. Then, add a new environment variable JAVA_HOME_V8 that point to the Java 1.8 home.